Map up-cased symbol names to referenced values if those values are string scalars (else throw).

Map up-cased symbol names to referenced values if those values are string scalars (else throw).

map_upper(...)

Arguments

...
symbol names mapping to string scalars

Value

map from original symbol names to new names (names found in the current environment)

See also

let

Examples

x <- 'a'
print(map_upper(x))

#> $X
#> [1] "a"
#> 

d <- data.frame(a = "a_val")
let(map_upper(x), paste(d$X, x))

#> [1] "a_val a"