Use function to reduce or expand arguments.

Use function f to reduce arguments ars (inspired by APL reduce/expand notation). This is essentially passing multiple values from a list as multiple independent arguments to a function. So this can be used to supply a standard evaluation interface for variadic functions. f %|.% args is roughly equivalent to do.call(f, args).

f %|.% args

args %.|% f

Arguments

f
function

args
argument list or vector

Value

f(args) where args elements become individual arguments of f.

Details

Note: these two operators are still experimental extensions. They will require more training material and use cases before they are fully advised.

Functions

%|.%: f reduce args

%.|%: args expand f

Examples

# basic examples
c(1, 2, 3) %.|% sum

#> [1] 6

c(1, 2, 3) %.|% base::sum

#> [1] 6

c(1, 2, 3) %.|% function(...) { sum(...) }

#> [1] 6

# partial application of log(5, base=2)
5 %.>% (c(., base=2) %.|% log)

#> [1] 2.321928

# # partial application with dplyr
# # can be used with dplyr/rlang as follows
# d <- data.frame(x=1, y=2, z=3)
# syms <- rlang::syms(c("x", "y"))
# d %.>% (c(list(.), syms) %.|% dplyr::select)