wrapr: Wrap R Tools for Debugging and Parametric Programming

wrapr: Wrap R Functions for Debugging and Parametric Programming

Provides DebugFnW() to capture function context on error for debugging, and let() which converts non-standard evaluation interfaces to parametric standard evaluation interfaces. DebugFnW() captures the calling function and arguments prior to the call causing the exception, while the classic options(error=dump.frames) form captures at the moment of the exception itself (thus function arguments may not be at their starting values). let() rebinds (possibly unbound) names to names.

For more information:

vignette('DebugFnW', package='wrapr')

vignette('let', package='wrapr')

vignette(package='wrapr')

Website: https://github.com/WinVector/wrapr

let video: https://youtu.be/iKLGxzzm9Hk?list=PLAKBwakacHbQp_Z66asDnjn-0qttTO-o9

Debug wrapper video: https://youtu.be/zFEC9-1XSN8?list=PLAKBwakacHbQT51nPHex1on3YNCCmggZA.

See also

Author

Maintainer: John Mount jmount@win-vector.com

Authors:

Nina Zumel nzumel@win-vector.com

Other contributors:

Win-Vector LLC [copyright holder]